add find all

+++ domain/model/time.go
@@ -1,7 +1,7 @@
 package model
 
 type TimeTrack struct {
-	CreatedAt int `json:"created_at" bson:"created_at"`
-	UpdatedAt int `json:"updated_at" bson:"updated_at"`
-	DeletedAt int `json:"deleted_at" bson:"deleted_at"`
+	CreatedAt int64 `json:"created_at" bson:"created_at"`
+	UpdatedAt int64 `json:"updated_at" bson:"updated_at"`
+	DeletedAt int64 `json:"deleted_at" bson:"deleted_at"`
 }
